bastanbenzer.sh
var1 is our n number, which indicates the lines to be checked.
var2 and var3 are the files to be compared.
If the line count or at least one of the files are less then our n number, the result will be "different" which means "farkli" in turkish.
If the first n lines (seen with head command) of file1 is not equal then its also "different". Else, it is the "same" which means "ayni".